An attempt at downgrading to Unity 2019, same version as DFU vanilla
This seems like it would work, but it requires using the recommended Android build tools for Unity 2019. These tools are no longer provided by Unity. They can still be installed, I think, but that would require sourcing them from the android sdk releases and manually installing them. As this downgrade was an attempt to *reduce* friction in modders supporting Android, I think it's best to drop the attempt and simply require modders to install Unity 2022 and keep the android fork as a separate project on their machine.