An attempt at making an object pool for UI Windows
It seems to kinda work, but it turns out this is a more difficult problem than expected. For it to truly work, we would need all windows to implement a Reset() and Initialize(params object[] args) method, where when they're returned to the pool they reset any problematic values (i.e. subscriptions to events, changes to their default values, etc), and when they're gotten from the pool they initialize any values given to the UIWindowFactory.GetInstance. I wouldn't mind doing the work to implement this for all DFU windows, but the problem is that it would break mod compatibility, for sure. Any custom window implementation made by a mod would have to implement the Reset and Initialize methods.