Added 'advanced options' to edit button and label anchors. Not yet saved.
Button anchor determines what corner of the screen it attaches to so that when you change aspect ratio it stays in the correct relative position. Label anchor determines if the button's label is on the top, bottom, etc. These settings are not yet saved in PlayerPrefs, so they revert when you close the app.